+++ tests/test_node.py
@@ -152,19 +152,6 @@ class TestNode(TestCase):
         self.assertIsNone(compare_tree)
         self.assertEqual(self, self)
 
-    def test_lca(self):
-        tree = self.tree
-        europe = tree['Europe']
-        result = Node.lca(europe['Norway'],
-                          europe.path.get("Sweden/Stockholm"),
-                          europe.path.get('Finland/Helsinki'))
-        self.assertEqual(europe, result)
-        result = Node.lca(europe,
-                          europe['Norway'],
-                          europe.path.get("Sweden/Stockholm"),
-                          europe.path.get('Finland/Helsinki'))
-        self.assertEqual(europe, result)
-
     def test_iter_together(self):
         tree = Node(identifier='world')
         tree['Europe'] = Node(data="first")
@@ -193,134 +180,6 @@ class TestNode(TestCase):
         self.assertNotEqual(tree, other_tree)
         self.assertNotEqual(other_tree, tree)
         self.assertEqual(tree, copy_tree)
-
-    def test_iter_children(self):
-        result = [str(child.path) for child in self.tree.children]
-        expected = ['/world/Europe', '/world/Africa']
-        self.assertEqual(expected, result)
-
-    def test_iter_tree1(self):
-        result = [str(child.path) for child in self.tree.iter_tree()]
-        expected = [
-            '/world',
-            '/world/Europe',
-            '/world/Europe/Norway',
-            '/world/Europe/Norway/Oslo',
-            '/world/Europe/Sweden',
-            '/world/Europe/Sweden/Stockholm',
-            '/world/Europe/Finland',
-            '/world/Europe/Finland/Helsinki',
-            '/world/Europe/Finland/Helsinki/Helsinki',
-            '/world/Europe/Finland/Helsinki/Helsinki/Helsinki',
-            '/world/Africa'
-        ]
-        self.assertEqual(expected, result)
-
-    def test_iter_tree2(self):
-        """Iterate only through nodes which are first-child."""
-        def is_first_child(_, item):
-            return item.index == 0
-
-        result = [str(child.path) for child in self.tree.iter_tree(is_first_child)]
-        expected = [
-            '/world',
-            '/world/Europe',
-            '/world/Europe/Norway',
-            '/world/Europe/Norway/Oslo',
-        ]
-        self.assertEqual(expected, result)
-
-    def test_iter_tree3(self):
-        target = self.tree
-        result = [child.identifier for child in target.iter_tree(keep=lambda _, it: it.depth < 3,
-                                                                 order="level")]
-        expected = ['world', 'Europe', 'Africa', 'Norway', 'Sweden', 'Finland']
-        self.assertEqual(expected, result)
-
-    def test_iter_ancestors(self):
-        target = self.tree.path(["Europe", "Norway", "Oslo"])
-        result = [str(child.path) for child in target.iter_ancestors()]
-        expected = ['/world/Europe/Norway', '/world/Europe', "/world"]
-        self.assertEqual(expected, result)
-
-    def test_iter_descendants1(self):
-        target = self.tree.path(["Europe", "Finland"])
-        result = [str(child.path) for child in target.iter_descendants(order="pre")]
-        expected = [
-            '/world/Europe/Finland/Helsinki',
-            '/world/Europe/Finland/Helsinki/Helsinki',
-            '/world/Europe/Finland/Helsinki/Helsinki/Helsinki',
-        ]
-        self.assertEqual(expected, result)
-
-    def test_iter_descendants_post1(self):
-        target = self.tree
-        result = [str(child.path) for child in target.iter_descendants(order="post")]
-
-        expected = [
-            '/world/Europe/Norway/Oslo',
-            '/world/Europe/Norway',
-            '/world/Europe/Sweden/Stockholm',
-            '/world/Europe/Sweden',
-            '/world/Europe/Finland/Helsinki/Helsinki/Helsinki',
-            '/world/Europe/Finland/Helsinki/Helsinki',
-            '/world/Europe/Finland/Helsinki',
-            '/world/Europe/Finland',
-            '/world/Europe',
-            '/world/Africa']
-        self.assertEqual(expected, result)
-
-    def test_iter_descendants_post2(self):
-        target = self.tree.path(["Europe", "Finland"])
-        result = [str(child.path) for child in target.iter_descendants(order="post")]
-        expected = [
-            '/world/Europe/Finland/Helsinki/Helsinki/Helsinki',
-            '/world/Europe/Finland/Helsinki/Helsinki',
-            '/world/Europe/Finland/Helsinki',
-        ]
-        self.assertEqual(expected, result)
-
-    def test_iter_descendants_post3(self):
-        def keep_square(_, item):
-            return item.index <= 2 and item.depth <= 2
-        target = self.tree
-        result = [str(child.path) for child in target.iter_descendants(keep_square, order="post")]
-        expected = [
-            '/world/Europe/Norway',
-            '/world/Europe/Sweden',
-            '/world/Europe/Finland',
-            '/world/Europe',
-            '/world/Africa',
-        ]
-        self.assertEqual(expected, result)
-
-    def test_iter_descendants3(self):
-        target = self.tree.path(["Europe", "Finland"])
-        result = [str(child.path) for child in target.iter_descendants(order="level")]
-        expected = [
-            '/world/Europe/Finland/Helsinki',
-            '/world/Europe/Finland/Helsinki/Helsinki',
-            '/world/Europe/Finland/Helsinki/Helsinki/Helsinki',
-        ]
-        self.assertEqual(expected, result)
-
-    def test_iter_siblings(self):
-        target = self.tree.path(["Europe", "Finland"])
-        result = [str(child.path) for child in target.iter_siblings()]
-        expected = ['/world/Europe/Norway', '/world/Europe/Sweden']
-        self.assertEqual(expected, result)
-
-    def test_iter_path(self):
-        target = self.tree.path(["Europe", "Norway", "Oslo"])
-        result = [str(node.path) for node in target.path]
-        expected = ['/world', '/world/Europe', '/world/Europe/Norway', '/world/Europe/Norway/Oslo']
-        self.assertEqual(expected, result)
-
-    def test_iter_leaves(self):
-        target = self.tree.path(["Europe"])
-        result = [child.identifier for child in target.iter_leaves()]
-        expected = ['Oslo', 'Stockholm', 'Helsinki']
-        self.assertEqual(expected, result)
 
     def test_copy(self):
         europe = self.tree["Europe"]
